requirements.txt
numpy>=1.24
pyrtlsdr>=0.3.0
pygame>=2.5.0
